+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### 🐞 Fixed
 - iOS 16 keyboard insets issue on pushed screen
+- Improved animation for date indicators in message list
 
 # [4.21.0](https://github.com/GetStream/stream-chat-swiftui/releases/tag/4.21.0)
 _September 02, 2022_